pytorch-stable-diffusion/requirements.txt
torch==2.5.1
numpy
tqdm
transformers
lightning
pillow
